+++ modules/sde-common/src/main/java/org/eclipse/tractusx/sde/common/entities/PolicyTemplateRequest.java
@@ -1,6 +1,6 @@
 /********************************************************************************
- * Copyright (c) 2022, 2024 T-Systems International GmbH
- * Copyright (c) 2022, 2024 Contributors to the Eclipse Foundation
+ * Copyright (c) 2024 T-Systems International GmbH
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -18,20 +18,26 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-package org.eclipse.tractusx.sde.common.exception;
+package org.eclipse.tractusx.sde.common.entities;
 
-import lombok.extern.slf4j.Slf4j;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
-@Slf4j
-public class ServiceException extends Exception {
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
 
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
+@Data
+@SuperBuilder
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+@EqualsAndHashCode(callSuper = true)
+public class PolicyTemplateRequest extends PolicyModel {
 
-	public ServiceException(String exceptionstr) {
-		super(exceptionstr);
-		log.info(exceptionstr);
-	}
+	@Builder.Default
+	private PolicyTemplateType type = PolicyTemplateType.NONE;
+
 }