+++ modules/sde-common/src/main/java/org/eclipse/tractusx/sde/common/entities/PolicyTemplateType.java
@@ -1,6 +1,6 @@
 /********************************************************************************
- * Copyright (c) 2022, 2024 T-Systems International GmbH
- * Copyright (c) 2022, 2024 Contributors to the Eclipse Foundation
+ * Copyright (c) 2024 T-Systems International GmbH
+ * Copyright (c) 2024 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -18,20 +18,8 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-package org.eclipse.tractusx.sde.common.exception;
+package org.eclipse.tractusx.sde.common.entities;
 
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public class ServiceException extends Exception {
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-
-	public ServiceException(String exceptionstr) {
-		super(exceptionstr);
-		log.info(exceptionstr);
-	}
+public enum PolicyTemplateType {
+	NONE, EXISTING, NEW
 }