+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
-## [Unreleased] - 2023-02-06
+## [Unreleased] - 2023-02-10
 
 ### Added
 
 - Policy-Hub service api integration.
+- Added New Policy Entities and pojo classes.
 
 ### Fixed